norminette fix

+++ libft/ft_strlcpy.c
@@ -1,13 +1,25 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strlcpy.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: khoubaib <khoubaib@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/10/07 21:12:18 by khoubaib          #+#    #+#             */
+/*   Updated: 2022/10/07 21:12:46 by khoubaib         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
-size_t ft_strlcpy(char *dst, const char *src, size_t size) 
+size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
-	int	i;
-	size_t slen;
+	size_t	i;
+	size_t	slen;
 
 	i = 0;
 	slen = ft_strlen(src);
-	if(size <= 0)
+	if (size <= 0)
 		return (slen);
 	while ((i < size - 1) && (src[i] != '\0'))
 	{
